fullstack-app/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.13
greenlet==3.2.3
h11==0.16.0
idna==3.10
numpy==2.2.6  # this version was needed for build_lambda.sh to complete successfully (was using 2.3.1 previously)
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
python-dotenv==1.1.0
sentry-sdk[fastapi]==2.26.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
# Authentication dependencies
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
fastapi-mail==1.4.1
slowapi==0.1.9
requests==2.32.3
# AWS Lambda adapter
mangum==0.17.0
